LIME-1748 - Moved Stub and VC Validation Steps to a Common folder structure
The background steps to navigate through the Core Stub to get to the passport Page, along with the VC Validation steps at the end of the scenarios
Have been moved to a common step-defs and page file
This will help split the page objects and methods we use directly on the Passport CRI page, and the ones we use outside of the Passport CRI Page
Removed Test files related to orch stub tests (model files, page files)
