+++ .secrets.baseline
@@ -133,84 +133,84 @@
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "b35d40d12423628d17dba54252182396a1d90103",
         "is_verified": false,
-        "line_number": 480
+        "line_number": 481
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "7d46dde65b8301c1eaf84f99441a92be55bcf5f1",
         "is_verified": false,
-        "line_number": 481
+        "line_number": 482
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "a24621f1e7f2aa3be1b3fd8aa158bb666fef7da0",
         "is_verified": false,
-        "line_number": 485
+        "line_number": 486
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "acdb3814fc474a0eff4b8a0c471dd3d588409995",
         "is_verified": false,
-        "line_number": 486
+        "line_number": 487
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "fe2031f80d50c6fe6902a22233162a77f5e72952",
         "is_verified": false,
-        "line_number": 490
+        "line_number": 491
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "bd51a1e663fbdda216f4a0af85e7889b1a7acae6",
         "is_verified": false,
-        "line_number": 492
+        "line_number": 493
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "4189dcfc39ca9e23b297ffbbae6f8fc2065e385f",
         "is_verified": false,
-        "line_number": 493
+        "line_number": 494
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "83d2570a5fba078cbc5d2d6167a3c95b096222aa",
         "is_verified": false,
-        "line_number": 498
+        "line_number": 499
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "3f80e3787a4e138a7d4dc0ded3f5523bdfe65af3",
         "is_verified": false,
-        "line_number": 499
+        "line_number": 500
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "7c186605d87b2294f2353966611a68619729a9f3",
         "is_verified": false,
-        "line_number": 503
+        "line_number": 504
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "180a11b0b86b1a47d9fd1caf3a38e217f8585958",
         "is_verified": false,
-        "line_number": 504
+        "line_number": 505
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportAPIPage.java",
         "hashed_secret": "c29bfc272e1012aacc9dcad023af303f34e8b08a",
         "is_verified": false,
-        "line_number": 505
+        "line_number": 506
       }
     ],
     "acceptance-tests/src/test/resources/axe.min.js": [
@@ -359,5 +359,5 @@
       }
     ]
   },
-  "generated_at": "2025-08-14T14:10:56Z"
+  "generated_at": "2025-08-27T14:12:36Z"
 }
