--- acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/model/AddressType.java
@@ -1,6 +1,0 @@
-package uk.gov.di.ipv.cri.passport.acceptance_tests.model;
-
-public enum AddressType {
-    CURRENT,
-    PREVIOUS
-}
