+++ acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/pages/PassportPageObject.java
@@ -1,11 +1,5 @@
 package uk.gov.di.ipv.cri.passport.acceptance_tests.pages;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.ObjectReader;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.junit.Assert;
@@ -19,75 +13,14 @@ import uk.gov.di.ipv.cri.passport.acceptance_tests.utilities.Driver;
 import uk.gov.di.ipv.cri.passport.acceptance_tests.utilities.TestDataCreator;
 import uk.gov.di.ipv.cri.passport.acceptance_tests.utilities.TestInput;
 
-import java.io.IOException;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
 
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.fail;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static uk.gov.di.ipv.cri.passport.acceptance_tests.pages.Headers.IPV_CORE_STUB;
 
 public class PassportPageObject extends UniversalSteps {
 
     private final ConfigurationService configurationService;
     private static final Logger LOGGER = LogManager.getLogger();
-
-    private static final String STUB_VC_PAGE_TITLE = "IPV Core Stub Credential Result - GOV.UK";
-
-    private static final String STUB_ERROR_PAGE_TITLE = "IPV Core Stub - GOV.UK";
-
-    // Should be separate stub page
-
-    @FindBy(xpath = "//*[@id=\"main-content\"]/p/a/button")
-    public WebElement visitCredentialIssuers;
-
-    @FindBy(xpath = "//*[@value=\"Passport CRI dev\"]")
-    public WebElement passportCRIDevLocalStub;
-
-    @FindBy(xpath = "//*[@value=\"Passport CRI Shared dev\"]")
-    public WebElement passportCRISharedDevLocalStub;
-
-    @FindBy(xpath = "//*[@value=\"Passport CRI Dev\"]")
-    public WebElement passportCRIDev;
-
-    @FindBy(xpath = "//*[@value=\"Passport CRI Shared Dev\"]")
-    public WebElement passportCRISharedDev;
-
-    @FindBy(xpath = "//*[@value=\"Passport CRI Build\"]")
-    public WebElement passportCRIBuild;
-
-    @FindBy(xpath = "//*[@value=\"Passport CRI Staging\"]")
-    public WebElement passportCRIStaging;
-
-    @FindBy(xpath = "//*[@value=\"Passport CRI Integration\"]")
-    public WebElement passportCRIIntegration;
-
-    @FindBy(id = "rowNumber")
-    public WebElement selectRow;
-
-    @FindBy(xpath = "//*[@id=\"main-content\"]/div/details/div/pre")
-    public WebElement JSONPayload;
-
-    @FindBy(xpath = "//*[@id=\"main-content\"]/div/details")
-    public WebElement errorResponse;
-
-    @FindBy(xpath = "//*[@id=\"main-content\"]/div/details/summary/span")
-    public WebElement viewResponse;
-
-    @FindBy(xpath = "//*[@id=\"main-content\"]/form[2]/div/button")
-    public WebElement searchButton;
-
-    @FindBy(xpath = "//*[@id=\"main-content\"]/form[2]/div/button")
-    public WebElement goToPassportCRIButton;
-
-    // ---------------------
 
     @FindBy(className = "error-summary")
     public WebElement errorSummary;
@@ -103,9 +36,6 @@ public class PassportPageObject extends UniversalSteps {
 
     @FindBy(xpath = "/html/body/div[2]/nav/ul/li[1]/a")
     public WebElement languageToggleWales;
-
-    @FindBy(id = "proveAnotherWayRadio")
-    public WebElement proveAnotherWayRadio;
 
     @FindBy(id = "govuk-notification-banner-title")
     public WebElement errorText;
@@ -247,146 +177,8 @@ public class PassportPageObject extends UniversalSteps {
         TestDataCreator.createDefaultResponses();
     }
 
-    // Should be in stub page
-    public void navigateToIPVCoreStub() {
-        Driver.get().manage().deleteAllCookies();
-
-        String coreStubUrl = configurationService.getCoreStubUrl(true);
-        Driver.get().get(coreStubUrl);
-        assertExpectedPage(IPV_CORE_STUB, false);
-    }
-
-    public void navigateToPassportCRIOnTestEnv() {
-        visitCredentialIssuers.click();
-        String passportCRITestEnvironment = configurationService.getPassportCRITestEnvironment();
-        LOGGER.info("passportCRITestEnvironment = " + passportCRITestEnvironment);
-
-        boolean sharedDev = passportCRITestEnvironment.toLowerCase().contains("shared");
-
-        boolean isUsingLocalStub = configurationService.isUsingLocalStub();
-        LOGGER.info("isUsingLocalStub = " + isUsingLocalStub);
-
-        if (isUsingLocalStub) {
-            // Local Stub - Passport CRI dev V1 or Passport CRI Shared dev V1
-            if (!sharedDev) {
-                passportCRIDevLocalStub.click();
-            } else {
-                passportCRISharedDevLocalStub.click();
-            }
-        } else if (passportCRITestEnvironment.toLowerCase().contains("dev")) {
-            // Hosted Stub - Passport CRI Dev V1 or Passport CRI Shared Dev V1
-            if (!sharedDev) {
-                passportCRIDev.click();
-            } else {
-                passportCRISharedDev.click();
-            }
-        } else if (passportCRITestEnvironment.toLowerCase().contains("build")) {
-            passportCRIBuild.click();
-        } else if (passportCRITestEnvironment.toLowerCase().contains("staging")) {
-            passportCRIStaging.click();
-        } else if (passportCRITestEnvironment.toLowerCase().contains("integration")) {
-            passportCRIIntegration.click();
-        } else {
-            LOGGER.info("No test environment is set");
-        }
-    }
-
-    public void searchForUATUser(String number) {
-        assertURLContains(
-                "credential-issuer?cri="
-                        + "passport-v1-cri-"
-                        + System.getenv("ENVIRONMENT").toLowerCase());
-        selectRow.sendKeys(number);
-        searchButton.click();
-    }
-
-    public void navigateToPassportResponse(String validOrInvalid) {
-        assertURLContains("callback");
-
-        if ("Invalid".equalsIgnoreCase(validOrInvalid)) {
-            assertExpectedPage(STUB_ERROR_PAGE_TITLE, true);
-            assertURLContains("callback");
-            BrowserUtils.waitForVisibility(errorResponse, 10);
-            errorResponse.click();
-        } else {
-            assertExpectedPage(STUB_VC_PAGE_TITLE, true);
-            assertURLContains("callback");
-            BrowserUtils.waitForVisibility(viewResponse, 10);
-            viewResponse.click();
-        }
-    }
-
-    public void navigateToPassportCRI() {
-        goToPassportCRIButton.click();
-    }
-
-    // ------------------
-
-    // Should be seperate page
-
     public void passportPageURLValidation(String path) {
         assertURLContains(path);
-    }
-
-    public void jsonErrorResponse(String expectedErrorDescription, String expectedErrorStatusCode)
-            throws JsonProcessingException {
-        String result = JSONPayload.getText();
-        LOGGER.info("result = " + result);
-
-        JsonNode insideError = getJsonNode(result, "errorObject");
-        LOGGER.info("insideError = " + insideError);
-
-        JsonNode errorDescription = insideError.get("description");
-        JsonNode statusCode = insideError.get("httpstatusCode");
-        String ActualErrorDescription = insideError.get("description").asText();
-        String ActualStatusCode = insideError.get("httpstatusCode").asText();
-
-        LOGGER.info("errorDescription = " + errorDescription);
-        LOGGER.info("statusCode = " + statusCode);
-        LOGGER.info("testErrorDescription = " + expectedErrorDescription);
-        LOGGER.info("testStatusCode = " + expectedErrorStatusCode);
-
-        Assert.assertEquals(expectedErrorDescription, ActualErrorDescription);
-        Assert.assertEquals(expectedErrorStatusCode, ActualStatusCode);
-    }
-
-    public void checkScoreInStubIs(String validityScore, String strengthScore) throws IOException {
-        scoreIs(validityScore, strengthScore, JSONPayload.getText());
-    }
-
-    public void scoreIs(String validityScore, String strengthScore, String jsonPayloadText)
-            throws IOException {
-        String result = jsonPayloadText;
-        LOGGER.info("result = " + result);
-        JsonNode vcNode = getJsonNode(result, "vc");
-        List<JsonNode> evidence = getListOfNodes(vcNode, "evidence");
-
-        String ValidityScore = evidence.get(0).get("validityScore").asText();
-        assertEquals(ValidityScore, validityScore);
-
-        String StrengthScore = evidence.get(0).get("strengthScore").asText();
-        assertEquals(StrengthScore, strengthScore);
-    }
-
-    public void assertCheckDetailsWithinVc(String checkDetailsType, String passportCriVc)
-            throws IOException {
-
-        JsonNode vcNode = getJsonNode(passportCriVc, "vc");
-        List<JsonNode> evidence = getListOfNodes(vcNode, "evidence");
-
-        String checkDetails = null;
-        if (checkDetailsType.equals("success")) {
-            checkDetails = evidence.get(0).get("checkDetails").toString();
-        } else {
-            checkDetails = evidence.get(0).get("failedCheckDetails").toString();
-        }
-        assertEquals("[{\"checkMethod\":\"data\"}]", checkDetails);
-    }
-
-    public void userNotFoundInThirdPartyErrorIsDisplayed() {
-        BrowserUtils.waitForVisibility(userNotFoundInThirdPartyBanner, 10);
-        Assert.assertTrue(userNotFoundInThirdPartyBanner.isDisplayed());
-        LOGGER.info(userNotFoundInThirdPartyBanner.getText());
     }
 
     public void userReEntersLastName(String invalidLastName) {
@@ -399,6 +191,7 @@ public class PassportPageObject extends UniversalSteps {
         FirstName.sendKeys(invalidFirstName);
     }
 
+    // this method is not currently used, saved for reuse in future
     public void userReEntersMiddleNames(String invalidMiddleNames) {
         MiddleNames.clear();
         MiddleNames.sendKeys(invalidMiddleNames);
@@ -454,7 +247,6 @@ public class PassportPageObject extends UniversalSteps {
         validToYear.sendKeys(passportSubject.getValidToYear());
     }
 
-    // Why is this invalid
     public void userEntersInvalidPassportDetails() {
         PassportPageObject passportPage = new PassportPageObject();
         passportPage.passportNumber.sendKeys("123456789");
@@ -466,10 +258,9 @@ public class PassportPageObject extends UniversalSteps {
         passportPage.validToDay.sendKeys("01");
         passportPage.validToMonth.sendKeys("01");
         passportPage.validToYear.sendKeys("2030");
-
-        //        BrowserUtils.waitForPageToLoad(10);
     }
 
+    // this method is not currently used, saved for reuse in future
     public void enterInvalidLastAndFirstName() {
         PassportPageObject passportPageObject = new PassportPageObject();
         passportPageObject.LastName.sendKeys("Parker!");
@@ -477,6 +268,7 @@ public class PassportPageObject extends UniversalSteps {
         passportPageObject.MiddleNames.sendKeys("@@@@@@@");
     }
 
+    // this method is not currently used, saved for reuse in future
     public void enterBirthYear(String day, String month, String year) {
         PassportPageObject passportPageObject = new PassportPageObject();
         passportPageObject.birthDay.clear();
@@ -490,6 +282,7 @@ public class PassportPageObject extends UniversalSteps {
         passportPageObject.birthYear.sendKeys(year);
     }
 
+    // this method is not currently used, saved for reuse in future
     public void enterValidToDate(String day, String month, String year) {
         PassportPageObject passportPageObject = new PassportPageObject();
         passportPageObject.validToDay.clear();
@@ -503,6 +296,7 @@ public class PassportPageObject extends UniversalSteps {
         passportPageObject.validToYear.sendKeys(year);
     }
 
+    // this method is not currently used, saved for reuse in future
     public void enterPassportNumber(String passportNumber) {
         PassportPageObject passportPage = new PassportPageObject();
         passportPage.passportNumber.clear();
@@ -538,10 +332,12 @@ public class PassportPageObject extends UniversalSteps {
         validToYear.sendKeys(passportSubject.getValidToYear());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertInvalidDoBInErrorSummary(String expectedText) {
         Assert.assertEquals(expectedText, InvalidDOBErrorInSummary.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertInvalidDoBOnField(String expectedText) {
         Assert.assertEquals(
                 expectedText, InvalidDateOfBirthFieldError.getText().trim().replace("\n", ""));
@@ -558,100 +354,91 @@ public class PassportPageObject extends UniversalSteps {
                 expectedText, InvalidValidToDateFieldError.getText().trim().replace("\n", ""));
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertInvalidPassportNumberInErrorSummary(String expectedText) {
         Assert.assertEquals(expectedText, InvalidPassportErrorInSummary.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertInvalidPassportNumberOnField(String expectedText) {
         Assert.assertEquals(
                 expectedText, PassportNumberFieldError.getText().trim().replace("\n", ""));
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertInvalidLastNameInErrorSummary(String expectedText) {
         Assert.assertEquals(expectedText, InvalidLastNameErrorInSummary.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertInvalidLastNameOnField(String expectedText) {
         Assert.assertEquals(
                 expectedText, InvalidLastNameFieldError.getText().trim().replace("\n", ""));
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertInvalidFirstNameInErrorSummary(String expectedText) {
         Assert.assertEquals(expectedText, InvalidFirstNameErrorInSummary.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertInvalidFirstNameOnField(String expectedText) {
         Assert.assertEquals(
                 expectedText, InvalidFirstNameFieldError.getText().trim().replace("\n", ""));
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertInvalidMiddleNameInErrorSummary(String expectedText) {
         Assert.assertEquals(expectedText, InvalidMiddleNamesErrorInSummary.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertInvalidMiddleNameOnField(String expectedText) {
         Assert.assertEquals(
                 expectedText, InvalidMiddleNamesFieldError.getText().trim().replace("\n", ""));
     }
 
-    public void ciInVC(String ci) throws IOException {
-        String result = JSONPayload.getText();
-        LOGGER.info("result = " + result);
-        JsonNode vcNode = getJsonNode(result, "vc");
-        JsonNode evidenceNode = vcNode.get("evidence");
-
-        List<String> cis = getCIsFromEvidence(evidenceNode);
-
-        if (StringUtils.isNotEmpty(ci)) {
-            if (cis.size() > 0) {
-                LOGGER.info("HELP " + Arrays.toString(cis.toArray()) + "    " + ci);
-                assertTrue(cis.contains(ci));
-            } else {
-                fail("No CIs found");
-            }
-        }
-    }
-
-    public void assertDocumentNumberInVc(String documentNumber) throws IOException {
-        String result = JSONPayload.getText();
-        LOGGER.info("result = " + result);
-        JsonNode vcNode = getJsonNode(result, "vc");
-        String passportNumber = getDocumentNumberFromVc(vcNode);
-        assertEquals(documentNumber, passportNumber);
-    }
-
+    // this method is not currently used, saved for reuse in future
     public void validateErrorPageHeading(String expectedText) {
         Assert.assertEquals(expectedText, pageHeader.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertPageHeading(String expectedText) {
         Assert.assertEquals(expectedText, pageHeader.getText().split("\n")[0]);
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertProveAnotherWayLinkText(String expectedText) {
         Assert.assertEquals(expectedText, getParent(proveAnotherWay).getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertErrorPrefix(String expectedText) {
         Assert.assertEquals(expectedText, errorText.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertFirstLineOfUserNotFoundText(String expectedText) {
         Assert.assertEquals(expectedText, userNotFoundInThirdPartyBanner.getText().split("\n")[0]);
     }
 
+    // this method is not currently used, saved for reuse in future
     public void youWillBeAbleToFindSentence(String expectedText) {
         Assert.assertEquals(expectedText, thereWasAProblemFirstSentence.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertPageSourceContains(String expectedText) {
         assert (Driver.get().getPageSource().contains(expectedText));
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertLastNameLabelText(String expectedText) {
         Assert.assertEquals(expectedText, getLabel(getParent(LastName)).getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertGivenNameLegendText(String expectedText) {
         Assert.assertEquals(
                 expectedText,
@@ -660,103 +447,85 @@ public class PassportPageObject extends UniversalSteps {
                         .getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertMiddleNameLabelText(String expectedText) {
         Assert.assertEquals(expectedText, getLabel(getParent(MiddleNames)).getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertGivenNameDescription(String expectedText) {
         Assert.assertEquals(
                 expectedText, getLabel(firstNameHint.findElement(By.xpath("./../.."))).getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertGivenNameHint(String expectedText) {
         Assert.assertEquals(expectedText, firstNameHint.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertMiddleNameHint(String expectedText) {
         Assert.assertEquals(expectedText, middleNameHint.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertDateOfBirthLegendText(String expectedText) {
         Assert.assertEquals(expectedText, dateOfBirthLegend.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertDateOfBirthHintText(String expectedText) {
         Assert.assertEquals(expectedText, dateOfBirthHint.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertBirthDayLabelText(String expectedText) {
         Assert.assertEquals(expectedText, getLabel(getParent(birthDay)).getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertBirthMonthLabelText(String expectedText) {
         Assert.assertEquals(expectedText, getLabel(getParent(birthMonth)).getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertBirthYearLabelText(String expectedText) {
         Assert.assertEquals(expectedText, getLabel(getParent(birthYear)).getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertValidToHintText(String expectedText) {
         Assert.assertEquals(expectedText, validToHint.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertPassportNumberLabelText(String expectedText) {
         Assert.assertEquals(expectedText, passportNumberFieldLabel.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertPassportNumberHintText(String expectedText) {
         Assert.assertEquals(expectedText, passportNumberHint.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertPageDescription(String expectedText) {
         Assert.assertEquals(expectedText, pageDescriptionHeading.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertValidToLegend(String expectedText) {
         Assert.assertEquals(expectedText, validToLegend.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertErrorSummaryText(String expectedText) {
         Assert.assertEquals(expectedText, errorSummaryTitle.getText());
     }
 
+    // this method is not currently used, saved for reuse in future
     public void assertCTATextAs(String expectedText) {
         assertEquals(Continue.getText(), expectedText);
-    }
-
-    private List<String> getCIsFromEvidence(JsonNode evidenceNode) throws IOException {
-        ObjectReader objectReader =
-                new ObjectMapper().readerFor(new TypeReference<List<JsonNode>>() {});
-        List<JsonNode> evidence = objectReader.readValue(evidenceNode);
-
-        List<String> cis =
-                getListOfNodes(evidence.get(0), "ci").stream()
-                        .map(JsonNode::asText)
-                        .collect(Collectors.toList());
-        return cis;
-    }
-
-    private JsonNode getJsonNode(String result, String vc) throws JsonProcessingException {
-        ObjectMapper objectMapper = new ObjectMapper();
-        JsonNode jsonNode = objectMapper.readTree(result);
-        return jsonNode.get(vc);
-    }
-
-    private String getDocumentNumberFromVc(JsonNode vcNode) throws IOException {
-        JsonNode credentialSubject = vcNode.findValue("credentialSubject");
-        List<JsonNode> evidence = getListOfNodes(credentialSubject, "passport");
-
-        String passportNumber = evidence.get(0).get("documentNumber").asText();
-        return passportNumber;
-    }
-
-    private List<JsonNode> getListOfNodes(JsonNode vcNode, String evidence) throws IOException {
-        JsonNode evidenceNode = vcNode.get(evidence);
-
-        ObjectReader objectReader =
-                new ObjectMapper().readerFor(new TypeReference<List<JsonNode>>() {});
-        return objectReader.readValue(evidenceNode);
     }
 
     private WebElement getParent(WebElement webElement) {
@@ -765,56 +534,6 @@ public class PassportPageObject extends UniversalSteps {
 
     private WebElement getLabel(WebElement webElement) {
         return webElement.findElement(By.tagName("label"));
-    }
-
-    private JsonNode getVCFromJson(String vc) throws JsonProcessingException {
-        String result = JSONPayload.getText();
-        LOGGER.info("result = " + result);
-        ObjectMapper objectMapper = new ObjectMapper();
-        JsonNode jsonNode = objectMapper.readTree(result);
-        return jsonNode.get(vc);
-    }
-
-    public void expiryAbsentFromVC(String exp) throws JsonProcessingException {
-        assertNbfIsRecentAndExpiryIsNull();
-    }
-
-    public void assertJtiIsPresentAndNotNull() throws JsonProcessingException {
-        String result = JSONPayload.getText();
-        LOGGER.info("result = " + result);
-        ObjectMapper objectMapper = new ObjectMapper();
-        JsonNode jsonNode = objectMapper.readTree(result);
-        JsonNode jtiNode = jsonNode.get("jti");
-        LOGGER.info("jti = " + jtiNode.asText());
-
-        assertNotNull(jtiNode.asText());
-    }
-
-    private void assertNbfIsRecentAndExpiryIsNull() throws JsonProcessingException {
-        String result = JSONPayload.getText();
-        LOGGER.info("result = " + result);
-        ObjectMapper objectMapper = new ObjectMapper();
-        JsonNode jsonNode = objectMapper.readTree(result);
-        JsonNode nbfNode = jsonNode.get("nbf");
-        JsonNode expNode = jsonNode.get("exp");
-        String nbf = jsonNode.get("nbf").asText();
-        LOGGER.info("nbf = " + nbfNode);
-        LOGGER.info("exp = " + expNode);
-        LocalDateTime nbfDateTime =
-                LocalDateTime.ofEpochSecond(Long.parseLong(nbf), 0, ZoneOffset.UTC);
-
-        assertNull(expNode);
-        assertTrue(isWithinRange(nbfDateTime));
-    }
-
-    boolean isWithinRange(LocalDateTime testDate) {
-        LocalDateTime nbfMin = LocalDateTime.now(ZoneOffset.UTC).minusSeconds(30);
-        LocalDateTime nbfMax = LocalDateTime.now(ZoneOffset.UTC).plusSeconds(30);
-        LOGGER.info("nbfMin " + nbfMin);
-        LOGGER.info("nbfMax " + nbfMax);
-        LOGGER.info("nbf " + testDate);
-
-        return testDate.isBefore(nbfMax) && testDate.isAfter(nbfMin);
     }
 
     private LocalDate subtractMonthsFromCurrentDate(int monthsToSubtract) {
@@ -828,6 +547,12 @@ public class PassportPageObject extends UniversalSteps {
                 pastDate);
 
         return pastDate;
+    }
+
+    public void userNotFoundInThirdPartyErrorIsDisplayed() {
+        BrowserUtils.waitForVisibility(userNotFoundInThirdPartyBanner, 10);
+        Assert.assertTrue(userNotFoundInThirdPartyBanner.isDisplayed());
+        LOGGER.info(userNotFoundInThirdPartyBanner.getText());
     }
 
     public void userReEntersExpiryDateAsCurrentDateMinus(int monthsToSubtract, int daysToSubtract) {
