+++ acceptance-tests/src/test/java/uk/gov/di/ipv/cri/passport/acceptance_tests/step_definitions/PassportStepDefs.java
@@ -1,14 +1,11 @@
 package uk.gov.di.ipv.cri.passport.acceptance_tests.step_definitions;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import uk.gov.di.ipv.cri.passport.acceptance_tests.pages.PassportPageObject;
 import uk.gov.di.ipv.cri.passport.acceptance_tests.utilities.BrowserUtils;
-
-import java.io.IOException;
 
 public class PassportStepDefs extends PassportPageObject {
 
@@ -62,16 +59,6 @@ public class PassportStepDefs extends PassportPageObject {
         userReEntersValidToYear(InvalidValidToYear);
     }
 
-    @Given("I navigate to the IPV Core Stub")
-    public void navigateToStub() {
-        navigateToIPVCoreStub();
-    }
-
-    @Then("^I search for passport user number (.*) in the Experian table$")
-    public void i_search_for_passport_user_number(String number) {
-        searchForUATUser(number);
-    }
-
     @And("I assert the url path contains (.*)$")
     public void i_assert_the_url_path_contains(String path) {
         passportPageURLValidation(path);
@@ -82,25 +69,9 @@ public class PassportStepDefs extends PassportPageObject {
         assertExpectedPage(pageTitle, false);
     }
 
-    @Then("^I navigate to the passport verifiable issuer to check for a (.*) response$")
-    public void i_navigate_to_passport_verifiable_issuer_for_valid_response(String validOrInvalid) {
-        navigateToPassportResponse(validOrInvalid);
-    }
-
-    @And("^JSON response should contain error description (.*) and status code as (.*)$")
-    public void errorInJsonResponse(String testErrorDescription, String testStatusCode)
-            throws JsonProcessingException {
-        jsonErrorResponse(testErrorDescription, testStatusCode);
-    }
-
     @Given("^I delete the (.*) cookie to get the unexpected error$")
     public void iDeleteTheCookieToGetTheUnexpectedError(String cookieName) {
         BrowserUtils.deleteCookie(cookieName);
-    }
-
-    @And("^I click the passport CRI for the testEnvironment$")
-    public void navigateToPassportOnTestEnv() {
-        navigateToPassportCRIOnTestEnv();
     }
 
     @When("^User Re-enters data as a (.*)$")
@@ -146,34 +117,6 @@ public class PassportStepDefs extends PassportPageObject {
     @When("User clicks on continue")
     public void user_clicks_on_continue() {
         Continue.click();
-    }
-
-    @And("^JSON payload should contain ci (.*), validity score (.*) and strength score (.*)$")
-    public void contraIndicatorInVerifiableCredential(
-            String ci, String validityScore, String strengthScore) throws IOException {
-        ciInVC(ci);
-        checkScoreInStubIs(validityScore, strengthScore);
-    }
-
-    @And("^JSON payload should contain validity score (.*) and strength score (.*)$")
-    public void scoresInVerifiableCredential(String validityScore, String strengthScore)
-            throws IOException {
-        checkScoreInStubIs(validityScore, strengthScore);
-    }
-
-    @And("^JSON response should contain documentNumber (.*) same as given passport$")
-    public void errorInJsonResponse(String documentNumber) throws IOException {
-        assertDocumentNumberInVc(documentNumber);
-    }
-
-    @And("^(.*) should be absent in the JSON payload$")
-    public void ExpiryNotPresentInJsonResponse(String exp) throws JsonProcessingException {
-        expiryAbsentFromVC(exp);
-    }
-
-    @And("^Passport VC should contain JTI field$")
-    public void jsonPayloadShouldContainJtiField() throws IOException {
-        assertJtiIsPresentAndNotNull();
     }
 
     @Then("User enters expiry date as current date minus (.*) months and minus (.*) days$")
