--- acceptance-tests/src/test/resources/Data/KennethDecerqueira.json
@@ -1,30 +1,0 @@
-{
-  "Q00002": "Q00002-2002-label",
-  "Q00009": "Q00009-SANTANDERANMFMORTGAGE-label",
-  "Q00015": "Q00015-OVER35000UPTO60000-label,Q00015-UPTO60000-label",
-  "Q00018": "Q00018-OVER500UPTO600-label,Q00018-UPTO600-label",
-  "Q00019": "Q00019-KA-label",
-  "Q00020": "Q00020-021963-label",
-  "Q00033": "Q00033-TSBBANKPLC-label",
-  "Q00036": "",
-  "Q00039": "Q00039-UPTO6750-label,Q00039-OVER6500UPTO6750-label",
-  "Q00040": "",
-  "Q00042": "Q00042-OVER550UPTO600-label,Q00042-UPTO600-label",
-  "Q00043": "Q00043-UPTO48MONTHS-label,Q00043-OVER36MONTHSUPTO48MONTHS-label",
-  "Q00048": "Q00048-TSBBANKPLC-label",
-  "Q00087": "",
-  "passportNumber": "321654987",
-  "surname": "Decerqueira",
-  "firstName": "Kenneth",
-  "middleName": "",
-  "dobDay": "08",
-  "dobMonth": "07",
-  "dobYear": "1965",
-  "expDay": "01",
-  "expMonth": "01",
-  "expYear": "2030",
-  "postcode": "BA2 5AA",
-  "address": "8 HADLEY ROAD, BATH, BA2 5AA",
-  "incorrectAddress": "6 HADLEY ROAD, BATH, BA2 5AA",
-  "yearMovedIn": "2000"
-}
